Adding auth again

+++ src/main/java/no/asgari/civilization/server/resource/AdminResource.java
@@ -65,14 +65,15 @@ public class AdminResource {
 
     /**
      * Since this will go in production, for now only I am allowed to change this
+     * @param admin
      * @param gameid
      */
     @Path("/deletegame")
     @POST
-    public Response deleteGame(@QueryParam("gameid") String gameid) {
-//        if(!admin.getUsername().equals("cash1981")) {
-//            return Response.status(Response.Status.FORBIDDEN).build();
-//        }
+    public Response deleteGame(@Auth Player admin, @QueryParam("gameid") String gameid) {
+        if(!admin.getUsername().equals("cash1981")) {
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
 
         boolean deleted = gameAction.deleteGame(gameid);
         if(deleted) return Response.ok().build();
